+++ test/integration/data/schema.sql
@@ -19,6 +19,7 @@ CREATE TABLE `users` (
   `country_id` int(11) DEFAULT NULL,
   PRIMARY KEY (`id`),
   UNIQUE KEY `unique_username` (`username`),
+  FULLTEXT KEY `fulltext_users_fields` (`username`, `first_name`, `last_name`),
   -- country
 	KEY `fk_users_country_id` (`country_id`),
   CONSTRAINT `fk_users_country_id` FOREIGN KEY (`country_id`) REFERENCES `country` (`id`)